+++ src/callbacksCustom/postNewRoomReordered.ts
@@ -12,7 +12,6 @@ import {
 import { updateCachedAPIFunctions } from "../cache";
 import { charCharOrderPostNewRoom } from "../features/changeCharOrder/callbacks/postNewRoom";
 import * as banFirstFloorRoomType from "../features/mandatory/banFirstFloorRoomType";
-import * as beastPreventEnd from "../features/mandatory/beastPreventEnd";
 import * as nerfCardReading from "../features/mandatory/nerfCardReading";
 import * as planetariumFix from "../features/mandatory/planetariumFix";
 import * as preventSacrificeRoomTeleport from "../features/mandatory/preventSacrificeRoomTeleport";
@@ -69,7 +68,6 @@ function main() {
   // Mandatory
   removeGloballyBannedItems.postNewRoom();
   nerfCardReading.postNewRoom();
-  beastPreventEnd.postNewRoom();
   tempMoreOptions.postNewRoom();
   banFirstFloorRoomType.postNewRoom();
   preventSacrificeRoomTeleport.postNewRoom();